requirements.txt
Django==1.9.5
django-after-response==0.2.1
django-debug-toolbar==1.4
django-extra-views==0.7.1
ecdsa==0.13
Fabric==1.11.1
jsonfield==1.0.3
paramiko==1.16.0
pycrypto==2.6.1
python-dateutil==2.5.3
python-wordpress-xmlrpc==2.3
requests==2.9.1
six==1.10.0
sqlparse==0.1.19
